fixed warning

+++ include/caffe/syncedmem.hpp
@@ -77,7 +77,7 @@ class SyncedMemory {
   void clear_data();
   void* cpu_ptr_;
   void* gpu_ptr_;
-  size_t size_ = -1;
+  int size_ = -1;
   SyncedHead head_;
   bool own_cpu_data_;
   bool cpu_malloc_use_cuda_;
